WebSecurity and DTOs Changes for adding JPA, added try catch for flows

+++ src/main/java/cfp/wecare/flow/ui/prgm/service/PrgmServiceImpl.java
@@ -1,17 +1,19 @@
 package cfp.wecare.flow.ui.prgm.service;
 
+import cfp.wecare.dto.OrgDto;
 import cfp.wecare.dto.PrgmDto;
 import cfp.wecare.flow.ui.prgm.Exception.PrgmException;
 import cfp.wecare.flow.ui.prgm.repository.PrgmRepository;
+import cfp.wecare.model.Org;
 import cfp.wecare.model.Prgm;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import javax.sound.midi.Soundbank;
-import java.util.List;
-import java.util.stream.Collectors;
+import java.util.*;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
 public class PrgmServiceImpl implements PrgmService {
@@ -23,24 +25,53 @@ public class PrgmServiceImpl implements PrgmService {
 
     @Override
     public List<PrgmDto> getPrgms() {
-        List<Prgm> prgms = (List<Prgm>) prgmRepository.findAll();
-        return prgms.stream()
-                .map(this::mapperToDto)
-                .collect(Collectors.toList());
+        try {
+            List<Prgm> prgms = (List<Prgm>) prgmRepository.findAll();
+            return prgms.stream()
+                    .filter(Objects::nonNull)
+                    .map(this::mapperToDto)
+                    .collect(toList());
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            throw new PrgmException(HttpStatus.INTERNAL_SERVER_ERROR, "Fetching Programs failed with internal error");
+        }
     }
 
     @Override
     public PrgmDto getPrgm(String prgmId) {
-        Prgm prgm = prgmRepository.findById(prgmId).orElseThrow(() -> new PrgmException(HttpStatus.NOT_FOUND, "Program not found"));
-        return mapperToDto(prgm);
+        try {
+            Optional<Prgm> prgm = prgmRepository.findById(prgmId);
+            if (prgm.isEmpty()) {
+                throw new PrgmException(HttpStatus.NOT_FOUND, "Program not found");
+            }
+            Prgm dbPrgm = prgm.get();
+            List<Org> orgs = dbPrgm.getOrgs();
+            List<OrgDto> orgDtos = new ArrayList<>();
+            for (Org org : orgs) {
+                orgDtos.add(modelMapper.map(org, OrgDto.class));
+            }
+            return PrgmDto.builder()
+                    .pgmId(dbPrgm.getPgmId())
+                    .pgmName(dbPrgm.getPgmName())
+                    .stDate(dbPrgm.getStDate())
+                    .edDate(dbPrgm.getEdDate())
+                    .orgs(orgDtos)
+                    .build();
+        } catch (PrgmException e) {
+            throw e;
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new PrgmException(HttpStatus.INTERNAL_SERVER_ERROR, "Fetching program details failed with internal error");
+        }
     }
 
     @Override
     public PrgmDto savePrgm(PrgmDto prgmDto) {
-        System.out.println("in save Prgm");
         try {
-            Prgm prgm = prgmRepository.save(modelMapper.map(prgmDto, Prgm.class));
-            return mapperToDto(prgm);
+            Prgm prgm = modelMapper.map(prgmDto, Prgm.class);
+            prgm.setPgmId(UUID.randomUUID().toString());
+            Prgm savedPrgm = prgmRepository.save(prgm);
+            return mapperToDto(savedPrgm);
         } catch (Exception e) {
             throw new PrgmException(HttpStatus.INTERNAL_SERVER_ERROR, "Could not create a program " + e.getMessage());
         }
@@ -49,9 +80,19 @@ public class PrgmServiceImpl implements PrgmService {
     @Override
     public PrgmDto updateProgram(String prgmId, PrgmDto prgmDto) {
         try {
-            Prgm prgm = modelMapper.map(prgmDto, Prgm.class);
-            prgm.setPgmId(prgmId);
-            return mapperToDto(prgmRepository.save(prgm));
+            Optional<Prgm> optionalPrgm = prgmRepository.findById(prgmId);
+            if (optionalPrgm.isEmpty()) {
+                throw new PrgmException(HttpStatus.NOT_FOUND, "The program does not exist");
+            }
+            Prgm prgm = optionalPrgm.get();
+            prgm.setPgmId(prgmDto.getPgmId());
+            prgm.setPgmName(prgmDto.getPgmName());
+            prgm.setStDate(prgmDto.getStDate());
+            prgm.setEdDate(prgmDto.getEdDate());
+            Prgm savedPrgm = prgmRepository.save(prgm);
+            return modelMapper.map(savedPrgm, PrgmDto.class);
+        } catch (PrgmException exception) {
+            throw exception;
         } catch (Exception e) {
             throw new PrgmException(HttpStatus.INTERNAL_SERVER_ERROR, "Could not update program " + e.getMessage());
         }
@@ -60,7 +101,12 @@ public class PrgmServiceImpl implements PrgmService {
     @Override
     public void deleteProgram(String prgmId) {
         try {
+            if (!prgmRepository.existsById(prgmId)) {
+                throw new PrgmException(HttpStatus.NOT_FOUND, "The program does not exist");
+            }
             prgmRepository.deleteById(prgmId);
+        } catch (PrgmException exception) {
+            throw exception;
         } catch (Exception e) {
             throw new PrgmException(HttpStatus.INTERNAL_SERVER_ERROR, "Could not delete program " + e.getMessage());
         }
